feat: add CustomLinkedList to the PerformanceBenchmark
